Add common error type for better error handling
Create new error type and interface to be used by EdgeX services for
more consistent error handling. The newly added errors are to be used by
any layer in the application and is expected to bubble up to the
transport(HTTP) layer which then the exported `ToHttpResponse` function
can be used to turn the error into an HTTP response. This will eliminate
the need for type checking errors at the HTTP layer. Also, these changes
are made expecting that services will not only return the proper
response code associated with the error, but also the error in the body
of the response. This will allow clients to unmarshal, and either log or
determine the correct course of action based on the type of error.

Signed-off-by: Anthony M. Bonafide <AnthonyMBonafide@gmail.com>